Refactor the source line mapping pass
This pass can be called with a command line option '-mapping-function',
defining the function to be mapped. The output includes the instruction
number, instruction dump, and the source file and line number.

If no mapping function is supplied, we then iterate all the functions of
the module, ignoring the debug and external function calls, of course.