Don't save the tid in the DynValue and DynBB
We can infer it from trace[index].tid. We previously added the thread id
because we're not sure we can get the trace[index].tid all the time,
e.g. when the index is not quite available, or the trace[] is not
visible in DynValue and DynBB. Actually if the index is not available,
the tid will not be available either. And we only use the tid at the
TraceFile.cpp where the trace[] is always visible.